Lambda function to archive GitHub events
GitHub provides a webhook for events which occur within an organisation. This
Lambda function can be placed behind API Gateway (Terraform yet to be written)
and used as a webhook for events occurring within an organisation.

Events are written to S3 in order to be used as an audit log of activity.