Validate the MAC provided by GitHub
Only archive events we have validated as originating from Github. The SHA1 HMAC
attached to each request as a header uses a shared secret which we pass in using
an environment variable.

We first implemented this using KMS to encrypt the shared secret but that
doesn't seem necessary for now.